Adjusting directory structure
