+++ app/src/androidTest/java/com/damienwesterman/defensedrill/database/DrillRepositoryTest.java
@@ -24,6 +24,11 @@ import android.database.sqlite.SQLiteConstraintException;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.platform.app.InstrumentationRegistry;
 
+import com.damienwesterman.defensedrill.data.CategoryEntity;
+import com.damienwesterman.defensedrill.data.Drill;
+import com.damienwesterman.defensedrill.data.DrillRepository;
+import com.damienwesterman.defensedrill.data.SubCategoryEntity;
+
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.Test;
