+++ app/src/main/AndroidManifest.xml
@@ -13,14 +13,14 @@
         android:theme="@style/Theme.DefenseDrill"
         tools:targetApi="31">
         <activity
-            android:name=".activities.CategorySelectActivity"
+            android:name=".ui.activities.CategorySelectActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".activities.HomeActivity"
+            android:name=".ui.activities.HomeActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
