+++ app/src/main/java/com/damienwesterman/defensedrill/data/DrillDatabase.java
@@ -9,7 +9,7 @@
  *                            *
  \****************************/
 
-package com.damienwesterman.defensedrill.database;
+package com.damienwesterman.defensedrill.data;
 
 import android.content.Context;
