+++ app/src/main/java/com/damienwesterman/defensedrill/data/DrillEntity.java
@@ -9,7 +9,7 @@
  *                            *
  \****************************/
 
-package com.damienwesterman.defensedrill.database;
+package com.damienwesterman.defensedrill.data;
 
 import androidx.annotation.NonNull;
 import androidx.room.ColumnInfo;
