+++ app/src/main/java/com/damienwesterman/defensedrill/data/DrillRepository.java
@@ -9,7 +9,7 @@
  *                            *
  \****************************/
 
-package com.damienwesterman.defensedrill.database;
+package com.damienwesterman.defensedrill.data;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteConstraintException;
