+++ app/src/main/res/layout/activity_category_select.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".activities.CategorySelectActivity">
+    tools:context=".ui.activities.CategorySelectActivity">
 
     <TextView
         android:id="@+id/textView"
