+++ app/src/main/res/layout/activity_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".activities.HomeActivity" >
+    tools:context=".ui.activities.HomeActivity" >
 
     <!-- Generate Drill card -->
     <androidx.cardview.widget.CardView
